Migrate MCP functionality to CLI with comprehensive workshop tools
