+++ MIGRATION_SUMMARY.md
@@ -1,0 +1,116 @@
+# Workshop MCP to CLI Migration Summary
+
+## Overview
+
+Successfully migrated all MCP (Model Context Protocol) tools, resources, and prompts from the Workshop MCP to the Workshop CLI. The MCP is now a simple wrapper around the CLI commands.
+
+## Changes Made
+
+### 1. Workshop CLI (`packages/workshop-cli/`)
+
+#### Dependencies Added
+- `@epic-web/invariant: ^1.0.0`
+- `openid-client: ^6.6.2`
+- `zod: ^3.25.71`
+
+#### New Files Created
+- `src/utils.ts` - Utility functions for workshop directory handling
+- `src/tools.ts` - Tool functions (login, logout, set-playground, update-progress)
+- `src/resources.ts` - Resource functions (workshop context, exercise context, diffs, user info)
+- `src/prompts.ts` - Prompt functions (quiz-me)
+
+#### CLI Commands Added
+- `epicshop login` - Login to the workshop
+- `epicshop logout` - Logout from the workshop
+- `epicshop set-playground` - Set playground environment
+- `epicshop update-progress` - Update lesson progress
+- `epicshop get-workshop-context` - Get workshop context
+- `epicshop get-exercise-context` - Get exercise context
+- `epicshop get-diff` - Get diff between apps
+- `epicshop get-progress-diff` - Get progress diff
+- `epicshop get-user-info` - Get user information
+- `epicshop get-user-access` - Get user access level
+- `epicshop get-user-progress` - Get user progress
+- `epicshop quiz-me` - Generate quiz for exercises
+
+#### Global Options Added
+- `--workshop-dir, -w` - Workshop directory path
+- `--format, -f` - Output format (json/pretty)
+
+### 2. Workshop MCP (`packages/workshop-mcp/`)
+
+#### Complete Rewrite
+- Removed all business logic files (`tools.ts`, `resources.ts`, `prompts.ts`, `utils.ts`)
+- Rewrote `index.ts` to be a thin wrapper around CLI commands
+- Uses `child_process.exec` to execute CLI commands
+
+#### Dependencies Removed
+- `@epic-web/invariant`
+- `@epic-web/workshop-utils`
+- `openid-client`
+
+#### Dependencies Added
+- `@epic-web/workshop-cli` (file:../workshop-cli)
+
+#### MCP Tools (now CLI wrappers)
+- `login` → `epicshop login`
+- `logout` → `epicshop logout`
+- `set_playground` → `epicshop set-playground`
+- `update_progress` → `epicshop update-progress`
+- `get_workshop_context` → `epicshop get-workshop-context`
+- `get_exercise_context` → `epicshop get-exercise-context`
+- `get_diff_between_apps` → `epicshop get-diff`
+- `get_exercise_step_progress_diff` → `epicshop get-progress-diff`
+- `get_user_info` → `epicshop get-user-info`
+- `get_user_access` → `epicshop get-user-access`
+- `get_user_progress` → `epicshop get-user-progress`
+
+#### MCP Prompts (now CLI wrappers)
+- `quiz_me` → `epicshop quiz-me`
+
+## Benefits
+
+1. **Single Source of Truth**: All workshop functionality now lives in the CLI
+2. **Consistency**: Same logic used whether accessing via CLI or MCP
+3. **Maintainability**: Only need to update CLI code, MCP automatically gets updates
+4. **Testability**: CLI commands can be tested independently
+5. **Flexibility**: Users can access all functionality directly via CLI
+6. **Reduced Complexity**: MCP is now much simpler (just a wrapper)
+
+## Usage Examples
+
+### CLI Usage
+```bash
+# Login to workshop
+epicshop login
+
+# Set playground to exercise 2, step 3, problem
+epicshop set-playground -e 2 -s 3 -t problem
+
+# Get workshop context as JSON
+epicshop get-workshop-context --format json
+
+# Get exercise context for exercise 3
+epicshop get-exercise-context -e 3
+
+# Get diff between two apps
+epicshop get-diff --app1 01.01.problem --app2 01.01.solution
+
+# Generate quiz for exercise 2
+epicshop quiz-me -e 2
+```
+
+### MCP Usage
+The MCP interface remains the same - all existing MCP tools and prompts continue to work exactly as before, but now they call the CLI under the hood.
+
+## Migration Complete
+
+✅ All MCP tools moved to CLI  
+✅ All MCP resources moved to CLI  
+✅ All MCP prompts moved to CLI  
+✅ MCP now wraps CLI commands  
+✅ Both packages build successfully  
+✅ All functionality preserved  
+✅ API compatibility maintained  
+
+The migration is complete and both packages are ready for use!
