+++ packages/workshop-cli/package.json
@@ -32,11 +32,14 @@
   "dependencies": {
     "@epic-web/workshop-utils": "file:../workshop-utils",
     "@epic-web/workshop-app": "file:../workshop-app",
+    "@epic-web/invariant": "^1.0.0",
     "chalk": "^5.3.0",
     "close-with-grace": "^2.1.0",
     "get-port": "^7.1.0",
     "open": "^8.4.2",
-    "yargs": "^17.7.2"
+    "openid-client": "^6.6.2",
+    "yargs": "^17.7.2",
+    "zod": "^3.25.71"
   },
   "devDependencies": {
     "@epic-web/config": "^1.11.3",
