+++ packages/workshop-cli/src/prompts.ts
@@ -1,15 +1,13 @@
 import { getExercises } from '@epic-web/workshop-utils/apps.server'
 import { getWorkshopConfig } from '@epic-web/workshop-utils/config.server'
-import { type McpServer } from '@modelcontextprotocol/sdk/server/mcp.js'
-import { type GetPromptResult } from '@modelcontextprotocol/sdk/types.js'
 import { z } from 'zod'
-import { exerciseContextResource } from './resources.js'
+import { getExerciseContext } from './resources.js'
 import {
 	handleWorkshopDirectory,
 	workshopDirectoryInputSchema,
 } from './utils.js'
 
-export const quizMeInputSchema = {
+export const quizMeSchema = z.object({
 	workshopDirectory: workshopDirectoryInputSchema,
 	exerciseNumber: z
 		.string()
@@ -17,15 +15,10 @@ export const quizMeInputSchema = {
 		.describe(
 			`The exercise number to get quizzed on (e.g., \`4\`). Leave blank for a random exercise.`,
 		),
-}
+})
 
-export async function quizMe({
-	workshopDirectory,
-	exerciseNumber: providedExerciseNumber,
-}: {
-	workshopDirectory: string
-	exerciseNumber?: string
-}): Promise<GetPromptResult> {
+export async function quizMe(input: z.infer<typeof quizMeSchema>) {
+	const { workshopDirectory, exerciseNumber: providedExerciseNumber } = input
 	const workshopRoot = await handleWorkshopDirectory(workshopDirectory)
 	const config = getWorkshopConfig()
 	let exerciseNumber = Number(providedExerciseNumber)
@@ -35,40 +28,25 @@ export async function quizMe({
 			exercises[Math.floor(Math.random() * exercises.length)]
 		exerciseNumber = randomExercise?.exerciseNumber ?? 0
 	}
-	return {
-		messages: [
-			{
-				role: 'user',
-				content: {
-					type: 'text',
-					text: `
+	
+	const exerciseContext = await getExerciseContext({
+		workshopDirectory,
+		exerciseNumber,
+	})
+
+	const prompt = `
 You are an expert teacher.
 
 Below is context about exercise ${exerciseNumber} in the workshop titled "${config.title}" (subtitled "${config.subtitle}") found at ${workshopRoot}.
 
 Please use this context to provide quiz questions, one at a time, to me to help me solidify my understanding of this material. Ask me the question, I will provide a response, you will either congratulate my correct understanding and move on or guide me to the correct answer with follow-up questions and hints until I either ask you to tell me the answer or ask you to continue to the next question.
-							`.trim(),
-				},
-			},
-			{
-				role: 'user',
-				content: {
-					type: 'resource',
-					resource: await exerciseContextResource.getResource({
-						workshopDirectory,
-						exerciseNumber,
-					}),
-				},
-			},
-		],
-	}
-}
 
-export function initPrompts(server: McpServer) {
-	server.prompt(
-		'quiz_me',
-		'Have the LLM quiz you on topics from the workshop exercises',
-		quizMeInputSchema,
-		quizMe,
-	)
+Exercise Context:
+${JSON.stringify(exerciseContext, null, 2)}
+	`.trim()
+
+	return {
+		prompt,
+		exerciseContext,
+	}
 }
