+++ packages/workshop-mcp/package.json
@@ -31,11 +31,9 @@
     "build:watch": "nx watch --projects=@epic-web/workshop-mcp -- nx run \\$NX_PROJECT_NAME:build"
   },
   "dependencies": {
-    "@epic-web/invariant": "^1.0.0",
-    "@epic-web/workshop-utils": "file:../workshop-utils",
+    "@epic-web/workshop-cli": "file:../workshop-cli",
     "@modelcontextprotocol/sdk": "^1.14.0",
     "@modelcontextprotocol/inspector": "^0.15.0",
-    "openid-client": "^6.6.2",
     "zod": "^3.25.71"
   },
   "devDependencies": {
